+++ CHANGELOG.md
@@ -3,3 +3,6 @@
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
+
+- Add browser note storage with YAML frontmatter serialization and config-based
+  adapter routing.